requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
pandas==2.2.3
openai==1.58.1
tiktoken==0.8.0

chromadb==0.5.23

langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-chroma==0.1.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4

langgraph==0.2.60

# Document loaders: PDF, DOCX, Excel
pypdf==4.0.2
docx2txt==0.8
openpyxl==3.1.5
unstructured==0.18.5

# Optional online search tool
tavily-python==0.5.0
